refactor(test-loop): silo legacy mock PMA + make peer_manager_sender optional (5/n)
**Verification:** existing tests run unchanged on the mock path; CI
green confirms correctness. This is a pure mechanical refactor — no
behavioral change.

Two prep changes for the T4 real-PMA wiring:

1. Move all legacy mock-PMA files into `setup/mock_pma/`:
   - `setup/peer_manager_actor.rs` → `setup/mock_pma/peer_manager_actor.rs`
   - `setup/drop_condition.rs` → `setup/mock_pma/drop_condition.rs`
   - Plus the supporting `state.rs`, `bridge.rs` files.

2. Make `peer_manager_sender` optional on `NodeExecutionData`:
   - Field type flips from
     `TestLoopActorHandle<TestLoopPeerManagerActor>` to
     `Option<TestLoopActorHandle<TestLoopPeerManagerActor>>`.
   - At T4.0 SHA, `setup_client` always populates `Some(...)` (mock
     path is the default in this PR); existing tests are unchanged.
   - T4 will introduce a real-PMA branch where this is `None`.
   - The bridge in `mock_pma/drop_condition_bridge.rs` `expect`s
     `Some` — at T4.0 SHA, every test still goes through the mock,
     so this `expect` cannot fire.

`is_chunk_validated_by`, `should_drop_chunk_by_height`,
`should_drop_chunk_for_protocol_upgrade` flipped from `fn` →
`pub(crate) fn` so the bridge can call them.

Stack: bases on T3 (#15652, "TestLoopTransport scaffolding").
