+++ test-loop-tests/src/setup/mod.rs
@@ -1,7 +1,7 @@
 pub mod builder;
 pub mod drop_condition;
 pub mod env;
-pub(crate) mod peer_manager_actor;
+pub(crate) mod mock_pma;
 pub(crate) mod rpc;
 pub mod setup;
 pub mod state;
