+++ test-loop-tests/src/tests/catching_up.rs
@@ -1,5 +1,5 @@
 use crate::setup::builder::TestLoopBuilder;
-use crate::setup::peer_manager_actor::HandlerResult;
+use crate::setup::mock_pma::HandlerResult;
 use crate::utils::rotating_validators_runner::RotatingValidatorsRunner;
 use crate::utils::transactions::get_anchor_hash;
 use itertools::Itertools as _;
@@ -139,7 +139,11 @@ fn test_catchup_random_single_part_sync_common(
         .build();
 
     for node_datas in &env.node_datas {
-        let peer_actor_handle = node_datas.peer_manager_sender.actor_handle();
+        let peer_actor_handle = node_datas
+            .legacy_mock_pma_sender
+            .as_ref()
+            .expect("test uses register_override_handler — must call .use_legacy_mock_pma()")
+            .actor_handle();
         let peer_actor = env.test_loop.data.get_mut(&peer_actor_handle);
         peer_actor.register_override_handler(Box::new(move |request| -> HandlerResult {
             if let NetworkRequests::PartialEncodedChunkMessage { partial_encoded_chunk, .. } =
@@ -339,7 +343,11 @@ fn slow_test_catchup_sanity_blocks_produced() {
             }
         };
 
-        let peer_actor_handle = node_datas.peer_manager_sender.actor_handle();
+        let peer_actor_handle = node_datas
+            .legacy_mock_pma_sender
+            .as_ref()
+            .expect("test uses register_override_handler — must call .use_legacy_mock_pma()")
+            .actor_handle();
         let peer_actor = env.test_loop.data.get_mut(&peer_actor_handle);
         peer_actor.register_override_handler(Box::new(move |request| -> HandlerResult {
             if let NetworkRequests::Block { block } = &request {
@@ -415,7 +423,11 @@ fn slow_test_all_chunks_accepted() {
     let seen_chunk_same_sender = Rc::new(RefCell::new(HashSet::<(AccountId, u64, ShardId)>::new()));
     for node_datas in &env.node_datas {
         let seen_chunk_same_sender = seen_chunk_same_sender.clone();
-        let peer_actor_handle = node_datas.peer_manager_sender.actor_handle();
+        let peer_actor_handle = node_datas
+            .legacy_mock_pma_sender
+            .as_ref()
+            .expect("test uses register_override_handler — must call .use_legacy_mock_pma()")
+            .actor_handle();
         let peer_actor = env.test_loop.data.get_mut(&peer_actor_handle);
         peer_actor.register_override_handler(Box::new(move |msg| -> HandlerResult {
             match msg {
