+++ test-loop-tests/src/tests/cross_shard_tx.rs
@@ -1,6 +1,6 @@
 use crate::setup::builder::TestLoopBuilder;
 use crate::setup::env::TestLoopEnv;
-use crate::setup::peer_manager_actor::HandlerResult;
+use crate::setup::mock_pma::HandlerResult;
 use crate::setup::state::NodeExecutionData;
 use crate::utils::get_node_data;
 use crate::utils::rotating_validators_runner::RotatingValidatorsRunner;
@@ -122,7 +122,11 @@ fn test_cross_shard_tx_common(Params { num_transfers, rotate_validators, drop_ch
 
     for node_datas in &env.node_datas {
         let rng = rng.clone();
-        let peer_actor_handle = node_datas.peer_manager_sender.actor_handle();
+        let peer_actor_handle = node_datas
+            .legacy_mock_pma_sender
+            .as_ref()
+            .expect("test uses register_override_handler — must call .use_legacy_mock_pma()")
+            .actor_handle();
         let peer_actor = env.test_loop.data.get_mut(&peer_actor_handle);
         peer_actor.register_override_handler(Box::new(move |request| -> HandlerResult {
             let mut rng = rng.write();
