+++ test-loop-tests/src/tests/optimistic_block.rs
@@ -1,7 +1,7 @@
 use crate::setup::builder::TestLoopBuilder;
 use crate::setup::drop_condition::DropCondition;
 use crate::setup::env::TestLoopEnv;
-use crate::setup::peer_manager_actor::HandlerResult;
+use crate::setup::mock_pma::HandlerResult;
 use itertools::Itertools;
 use near_async::time::Duration;
 use near_chain_configs::test_genesis::{TestEpochConfigBuilder, ValidatorsSpec};
@@ -242,7 +242,13 @@ fn alter_optimistic_block_at_height(
     use near_primitives::optimistic_block;
 
     for data in &env.node_datas {
-        let peer_actor = env.test_loop.data.get_mut(&data.peer_manager_sender.actor_handle());
+        let peer_actor = env.test_loop.data.get_mut(
+            &data
+                .legacy_mock_pma_sender
+                .as_ref()
+                .expect("test uses register_override_handler — must call .use_legacy_mock_pma()")
+                .actor_handle(),
+        );
         peer_actor.register_override_handler(Box::new({
             let validator_signer = signer.clone();
             move |request: NetworkRequests| {
