+++ test-loop-tests/src/tests/spice.rs
@@ -1,7 +1,7 @@
 use super::spice_utils::delay_endorsements_propagation;
 use crate::setup::builder::TestLoopBuilder;
 use crate::setup::env::TestLoopEnv;
-use crate::setup::peer_manager_actor::HandlerResult;
+use crate::setup::mock_pma::HandlerResult;
 use crate::utils::account::{
     create_account_id, create_validators_spec, validators_spec_clients,
     validators_spec_clients_with_rpc,
@@ -824,7 +824,13 @@ fn test_spice_validator_only_does_not_distribute_witness_and_receipts() {
     for i in num_producers..env.node_datas.len() {
         let node_data = &env.node_datas[i];
         let counter = spice_data_sent_count.clone();
-        let peer_actor = env.test_loop.data.get_mut(&node_data.peer_manager_sender.actor_handle());
+        let peer_actor = env.test_loop.data.get_mut(
+            &node_data
+                .legacy_mock_pma_sender
+                .as_ref()
+                .expect("test uses register_override_handler — must call .use_legacy_mock_pma()")
+                .actor_handle(),
+        );
         peer_actor.register_override_handler(Box::new(move |request| {
             if matches!(&request, NetworkRequests::SpicePartialData { .. }) {
                 counter.fetch_add(1, Ordering::SeqCst);
@@ -868,7 +874,13 @@ fn test_spice_validator_only_sends_endorsements() {
     for i in num_producers..env.node_datas.len() {
         let node_data = &env.node_datas[i];
         let counter = endorsement_count.clone();
-        let peer_actor = env.test_loop.data.get_mut(&node_data.peer_manager_sender.actor_handle());
+        let peer_actor = env.test_loop.data.get_mut(
+            &node_data
+                .legacy_mock_pma_sender
+                .as_ref()
+                .expect("test uses register_override_handler — must call .use_legacy_mock_pma()")
+                .actor_handle(),
+        );
         peer_actor.register_override_handler(Box::new(move |request| {
             if matches!(&request, NetworkRequests::SpiceChunkEndorsement(..)) {
                 counter.fetch_add(1, Ordering::SeqCst);
